@@ -17,95 +15,69 @@
   "author": "Rodrick Alcantara <rialcantara21@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "dev": "react-scripts start",
+    "test": "jest",
+    "dev": "vite",
     "clean": "rm -rf ./build",
-    "build": "yarn clean && react-scripts build && compress-cra",
-    "start": "yarn build && yarn serve -s ./build",
-    "test": "react-scripts test",
-    "analyze": "yarn build && source-map-explorer ./build/static/js/*.js",
+    "build": "yarn clean && tsc && vite build",
+    "preview": "vite preview",
     "format": "prettier --write ./src/**/*.ts{,x}",
     "lint": "eslint ./src --ext .js,.ts,.jsx,.tsx",
     "lint:fix": "eslint ./src --fix --ext .js,.ts,.jsx,.tsx",
-    "git:develop": "git checkout develop && git pull origin develop && git branch | grep -v \"main\\|staging\\|develop\" | xargs git branch -D",
-    "storybook": "start-storybook -p 6006 -s ./build",
-    "build:storybook": "build-storybook -s ./build",
     "pre-commit": "lint-staged --allow-empty",
     "prepare": "husky install"
   },
-  "resolutions": {
-    "react-error-overlay": "6.0.9"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "axios": "^0.27.2",
+    "date-fns": "^2.28.0",
+    "immer": "^9.0.14",
+    "lodash": "^4.17.21",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-use": "^17.4.0",
+    "styled-components": "^5.3.5",
+    "yup": "^0.32.11",
+    "zustand": "^4.0.0-rc.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^15.0.0",
-    "@commitlint/config-conventional": "^15.0.0",
-    "@storybook/addon-actions": "^6.4.0",
-    "@storybook/addon-essentials": "^6.4.0",
-    "@storybook/addon-links": "^6.4.0",
-    "@storybook/node-logger": "^6.4.0",
-    "@storybook/preset-create-react-app": "^3.2.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@commitlint/cli": "^17.0.0",
+    "@commitlint/config-conventional": "^17.0.0",
+    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
     "@types/dotenv": "^8.2.0",
-    "@types/jest": "^26.0.15",
-    "@types/lodash": "^4.14.177",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/redux-logger": "^3.0.9",
-    "@types/styled-components": "^5.1.15",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "babel-loader": "8.1.0",
-    "compress-create-react-app": "^1.2.1",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "19.0.1",
-    "eslint-config-airbnb-typescript": "^16.0.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "2.25.3",
-    "eslint-plugin-jest": "^25.3.0",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/styled-components": "^5.1.25",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "@vitejs/plugin-react": "^1.3.0",
+    "dotenv": "^16.0.1",
+    "eslint": "8.2.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.2.2",
     "eslint-plugin-jsx-a11y": "6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "7.27.1",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "eslint-plugin-storybook": "^0.5.0",
-    "husky": "^6.0.0",
-    "lint-staged": "^12.1.2",
-    "node-sass": "^6.0.1",
-    "prettier": "^2.5.1",
-    "react-error-overlay": "6.0.9",
-    "typescript": "^4.1.2"
-  },
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.7.1",
-    "@storybook/react": "^6.4.0",
-    "axios": "^0.25.0",
-    "date-fns": "^2.28.0",
-    "dotenv": "^15.0.0",
-    "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.0.2",
-    "react-scripts": "4.0.3",
-    "react-use": "^17.3.2",
-    "redux-logger": "^3.0.6",
-    "styled-components": "^5.3.3",
-    "web-vitals": "^1.0.1",
-    "yup": "^0.32.11"
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "husky": "6",
+    "jest": "^28.1.0",
+    "lint-staged": "^12.4.1",
+    "node-sass": "^7.0.1",
+    "prettier": "^2.6.2",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.9",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-ejs": "^1.4.4",
+    "vite-plugin-svgr": "^2.1.0",
+    "vite-tsconfig-paths": "^3.4.1"
   },
   "lint-staged": {
     "*.{js,ts,jsx,tsx}": "eslint --cache --fix",
