@@ -4,15 +4,16 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAsyncFn } from 'react-use';
 import { ROUTES } from 'shared/constants/Routes';
-import UserActions from 'shared/redux/User/Actions';
+import useAppStore from 'shared/store';
 import CommonUtil from 'shared/utils/Common';
 
 const Home: React.FC = () => {
   const navigate = useNavigate();
-  const { logout } = UserActions();
+  const logout = useAppStore(state => state.logout);
 
+  // eslint-disable-next-line @typescript-eslint/require-await
   const [, logoutAsync] = useAsyncFn(async () => {
-    const { error } = await logout();
+    const { error } = logout();
 
     if (!error) return navigate(ROUTES.LOGIN);
 
