@@ -2,7 +2,7 @@ import { LayoutWrapper } from './Layout.styled';
 
 import React, { Suspense } from 'react';
 
-const Layout: React.FC = ({ children }) => {
+const Layout: React.FC<{ children: any }> = ({ children }) => {
   return (
     <LayoutWrapper>
       <Suspense fallback={<div>Loading...</div>}>
