@@ -7,12 +7,13 @@ import { ROUTES } from 'shared/constants/Routes';
 import { useForm } from 'shared/hooks/useForm';
 import { ILoginForm } from 'shared/interfaces/Auth';
 import { IFormElements } from 'shared/interfaces/Form';
-import UserActions from 'shared/redux/User/Actions';
+import useAppStore from 'shared/store';
 import CommonUtil from 'shared/utils/Common';
 import { loginValidator } from 'shared/validators/Login';
 
 const Login: React.FC = () => {
   const navigate = useNavigate();
+  const login = useAppStore(state => state.login);
 
   const defaultValues: ILoginForm = {
     email: '',
@@ -30,12 +31,11 @@ const Login: React.FC = () => {
     handle({ event });
   });
 
-  const { login } = UserActions();
-
+  // eslint-disable-next-line @typescript-eslint/require-await
   const [, loginAsync] = useAsyncFn(async (...args: [ILoginForm]) => {
-    const [form] = args;
+    const [] = args;
 
-    const { error } = await login(form);
+    const { error } = login();
 
     if (!error) return navigate(ROUTES.HOME);
 
