@@ -1,6 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
 // jest-dom adds custom jest matchers for asserting on DOM nodes.
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-// eslint-disable-next-line import/no-extraneous-dependencies
 import '@testing-library/jest-dom';
