@@ -1,5 +1,1 @@
-import dotenv from 'dotenv';
-dotenv.config();
-
-/* APP */
-export const APP_ZONE = process.env.REACT_APP_APP_ZONE || '';
+export const APP_ZONE = process.env.APP_ZONE || '';
